[WIP] DO NOT MERGE: Wconversion
Up for a discussion. To turn on Wconversion to warn on precision
loss of implicit conversion (would have caught a bug in the x86-64
assembler), quite a number of changes are necessary. size_t is prolific,
but the mirror objects all take 32b values. Dex things are often 16b.
Etc.

This CL can compile the first 15 or so .cc files and is by far not
complete. It also turns of sign-conversion warnings, so that the
prolific mismatch of int32_t and uint32_t doesn't matter.

+++ patchoat/patchoat.cc
@@ -728,7 +728,8 @@ bool PatchOat::PatchTextSection(ElfFileImpl* oat_file) {
   auto oat_text_sec = oat_file->FindSectionByName(".text");
   CHECK(oat_text_sec != nullptr);
   uint8_t* to_patch = oat_file->Begin() + oat_text_sec->sh_offset;
-  uintptr_t to_patch_end = reinterpret_cast<uintptr_t>(to_patch) + oat_text_sec->sh_size;
+  uintptr_t to_patch_end = reinterpret_cast<uintptr_t>(to_patch) +
+      static_cast<uintptr_t>(oat_text_sec->sh_size);
 
   for (; patches < patches_end; patches++) {
     CHECK_LT(*patches, oat_text_sec->sh_size) << "Bad Patch";
